+++ src/pages/Restrictions/contexts/mapLayersReducer.ts
@@ -1,6 +1,7 @@
 export const vehicleAxleWeightLayerId = 'vehicleAxleWeight'
 export const vehicleHeightLayerId = 'vehicleHeight'
 export const vehicleLengthLayerId = 'vehicleLength'
+export const vehicleMaxAllowedWeightLayerId = 'vehicleMaxAllowedWeight'
 export const vehicleWeightLayerId = 'vehicleWeight'
 export const vehicleWidthLayerId = 'vehicleWidth'
 
@@ -8,6 +9,7 @@ export const mapLayersInitialState = {
   [vehicleAxleWeightLayerId]: true,
   [vehicleHeightLayerId]: false,
   [vehicleLengthLayerId]: false,
+  [vehicleMaxAllowedWeightLayerId]: false,
   [vehicleWeightLayerId]: false,
   [vehicleWidthLayerId]: false,
 }
@@ -16,6 +18,7 @@ export const layerIds = [
   'vehicleAxleWeight',
   'vehicleHeight',
   'vehicleLength',
+  'vehicleMaxAllowedWeight',
   'vehicleWeight',
   'vehicleWidth',
 ] as const